Program færdig men kan finpudses stadig

+++ src/main/java/Controller/Bruger.java
@@ -11,6 +11,7 @@ import javax.ws.rs.core.MediaType;
 import java.sql.SQLException;
 
 import java.util.HashMap;
+import java.util.List;
 
 @Path("/bruger")
 public class Bruger {
@@ -27,8 +28,8 @@ public class Bruger {
     @GET
     public String getBrugerer() {
         try {
-            //UserDTO user = iUserDAO.getUsers();
-            return "kat";
+            List<UserDTO> users = iUserDAO.getUserList();
+            return Mapper.mapUsersDTO(users);
         } catch (Exception e){
             e.printStackTrace();
         }
@@ -40,7 +41,7 @@ public class Bruger {
     public String getBruger(@PathParam("username") String username) {
         try {
             UserDTO user = iUserDAO.getUser(username);
-            return "Du FANDT: " + ((user != null) ? user : "Ingen bruger");
+            return "Du FANDT: " + ((user != null) ? user : "INGEN BRUGER");
         } catch (Exception e){
             e.printStackTrace();
         }
@@ -51,6 +52,9 @@ public class Bruger {
     public String createUser(String JSON_userDTO) {
         try{
             UserDTO userDTO = Mapper.mapUserDTO(JSON_userDTO);
+            if(iUserDAO.getUser(userDTO.getUserName()) != null){
+                return "Brugeren " + userDTO.getUserName() + " eksisterer allerede!";
+            }
             userDTO.generateUserId();
             iUserDAO.createUser(userDTO);
             return "Brugeren " + userDTO.getUserName() + " blev oprettet med success!";
@@ -66,6 +70,9 @@ public class Bruger {
         try{
             UserDTO userDTO = Mapper.mapUserDTO(JSON_userDTO);
             UserDTO oldUserDTO = iUserDAO.getUser(username);
+            if(oldUserDTO == null){
+                return "Brugeren " + username + " eksisterer ikke!";
+            }
             userDTO.setUserId(oldUserDTO.getUserId());
             iUserDAO.updateUser(userDTO);
             return "Brugeren " + username + " blev redigeret med success!";
@@ -80,6 +87,9 @@ public class Bruger {
     public String deleteUser(@PathParam("username") String username) {
         try{
             UserDTO oldUserDTO = iUserDAO.getUser(username);
+            if(oldUserDTO == null){
+                return "Brugeren " + username + " eksisterer ikke!";
+            }
             iUserDAO.deleteUser(oldUserDTO.getUserId());
             return "Brugeren " + oldUserDTO.getUserName() + " blev slettet med success!";
         }
